requirements.txt
flask 
flask-smorest 
python-dotenv 
bcrypt 
flask-sqlalchemy 
flask-migrate
flask-cors
flask-jwt-extended
pymysql